+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.7] - 2022-01-26
+
+### Added
+
+- Created default parameters for each exotype
+
+### Changed
+
+- Refactored the system builder
+
 ## [1.0.6] - 2022-01-19
 
 ### Added